requirements.txt
openai==1.3.0
python-dotenv==1.0.0
rich==13.7.0
httpx==0.24.1 